Added pause in Restore-WinGetSoftware
When the cmdlet is not running as an admin, a warning is displayed.
A pause has been introduced to ensure the user sees this warning.
This may be skipped via the $Force switch which skips other forms
of user confirmation.

+++ WinGet-Essentials/modules/WinGet-Restore.psm1
@@ -98,8 +98,10 @@ function Restore-WinGetSoftware
         return
     }
 
-    if (-not(Test-Administrator)) {
-        Write-Warning "Some programs will not install correctly if winget is used without administrator rights. This is particularly true for zip-based installs."
+    if (-not(Test-Administrator) -and -not($Force)) {
+        Write-Warning 'Some programs will not install correctly if winget is used without administrator rights. This is particularly true for zip-based installs.'
+        Write-Host -NoNewLine 'Press any key to continue ...';
+        $null = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown');
     }
 
     if ($Force -and -not $Confirm){
